+++ src/main/java/edu/umn/ecology/populus/model/sspg/SSPGTable.java
@@ -54,14 +54,14 @@ public class SSPGTable extends AbstractTableModel {
         temp[dimension - 1 + 3] = new Vector(dimension);
         table = temp;
 
-        table[0].add(Integer.valueOf(dimension));
+        table[0].add(dimension);
         table[1].add(title);
-        table[2].add(Double.valueOf(initPop));
+        table[2].add(initPop);
 
         for (int i = 0; i < dimension - 1; i++)
-            table[3 + dimension - 1].add(Double.valueOf(0));
+            table[3 + dimension - 1].add((double) 0);
         for (int i = 0; i < dimension; i++)
-            table[3 + i].add(Double.valueOf(0));
+            table[3 + i].add((double) 0);
 
         this.fireTableStructureChanged();
     }
@@ -158,7 +158,7 @@ public class SSPGTable extends AbstractTableModel {
         double[][] data = new double[dimension][dimension];
         for (int i = 0; i < data.length; i++)
             for (int j = 0; j < data[i].length; j++)
-                data[i][j] = ((Double) getValueAt(i, j + 3)).doubleValue();
+                data[i][j] = (Double) getValueAt(i, j + 3);
         return data;
     }
 
@@ -188,7 +188,7 @@ public class SSPGTable extends AbstractTableModel {
                 this.setValueAt(mat[i][j], i, j + 3);
 
         Vector p = new Vector();
-        for (double pop : pops) p.add(Double.valueOf(pop));
+        for (double pop : pops) p.add(pop);
         table[2] = p;
 
         p = new Vector();
